Updating Documentation
Added comments to the Dockerfiles explaining the function of each code segment.
Added a few comments to the Args Class to clarify the changes from the original algorithm.
Updated a typo in the Code_Overview PDF file.
Added a slide to the Reliable Modular System PowerPoint.
Removed two libraries from requirements.txt that are not used.